fixed #65

+++ src/main/java/org/woehlke/simpleworklist/control/UserMessageController.java
@@ -30,7 +30,7 @@ public class UserMessageController extends AbstractController {
         UserAccount sender = super.userService.findUserById(userId);
         List<UserMessage> userMessageList = userMessageService.getLast20MessagesBetweenCurrentAndOtherUser(receiver,sender);
         for(UserMessage userMessage:userMessageList){
-            if((!userMessage.isReadByReceiver()) && (receiver.getId()==userMessage.getReceiver().getId())){
+            if((!userMessage.isReadByReceiver()) && (receiver.getId().longValue()==userMessage.getReceiver().getId().longValue())){
                 userMessage.setReadByReceiver(true);
                 userMessageService.update(userMessage);
             }
